balance calculation

+++ BudgetUI/Registration.xaml.cs
@@ -32,7 +32,7 @@ namespace BudgetUI
             {
                 if (PasswordBox_password.Password.Length > 0)
                 {
-                    if (Int32.TryParse(TextBox_budget.Text, out int budget))
+                    if (Decimal.TryParse(TextBox_budget.Text, out decimal budget))
                     {
                         //calculations.AddUser(TextBox_name.Text, PasswordBox_password.Password, budget);
                         Login lw = new Login();
